Remove MythRec and IPlayerData, fix mongo to not overflow the stack generating a huge warning, fix missed camelCase issues, add parseTime to rec parser
