+++ src/db/mongo/model/RecordedGameModel.ts
@@ -1,14 +1,7 @@
 import mongoose, { model } from "mongoose";
 import RecordedGameSchema from "../schemas/recordedGame/RecordedGameSchema";
-import { RecordedGameMetadata } from "@/types/RecordedGame";
+import { IRecordedGame } from "@/types/RecordedGame";
 
-const RecordedGameModel = mongoose.models.RecordedGames || model("RecordedGames", RecordedGameSchema);
-
-export interface IRecordedGame extends RecordedGameMetadata {
-  uploadedBy: string;
-  gameTitle: string;
-  createdAt: Date;
-  downloadCount: number;
-}
+const RecordedGameModel: mongoose.Model<IRecordedGame> = mongoose.models.RecordedGames || model<IRecordedGame>("RecordedGames", RecordedGameSchema);
 
 export default RecordedGameModel;
