+++ src/server/controllers/mongo-controller.ts
@@ -1,7 +1,7 @@
-import { MythRec } from "@/types/MythRecs";
 import { queryMythRecs } from "../services/mongo-service";
+import { IRecordedGame } from "@/types/RecordedGame";
 
 // todo - add pagination params
-export async function getMythRecs(page: number): Promise<MythRec[]> {
+export async function getMythRecs(page: number): Promise<IRecordedGame[]> {
   return await queryMythRecs(page);
 }
