+++ src/server/controllers/upload-rec-controller.ts
@@ -1,5 +1,5 @@
 import { parseRecordedGameMetadata } from "../recParser";
-import { RecordedGameMetadata } from "@/types/RecordedGame";
+import { RecordedGameMetadata } from "@/types/RecordedGameParser";
 import getMongoClient from "@/db/mongo/mongo-client";
 import RecordedGameModel from "@/db/mongo/model/RecordedGameModel";
 import { uploadRecToS3 } from "../services/aws";
@@ -44,6 +44,4 @@ export default async function uploadRec(params: UploadRecParams): Promise<void> 
     RecordedGameModel.deleteOne({ gameguid: recGameMetadata.gameGuid });
     throw new Error("Error uploading to s3");
   }
-
-  return result;
 }
