+++ src/utils/utils.ts
@@ -23,11 +23,10 @@ export function recMetadataSchemaHelper<T extends (...args: any) => any>(
   return obj;
 }
 
-export function removeMongoObjectID<T>(records: Array<T & {_id?: mongoose.Types.ObjectId}>): T[]
+export function removeMongoObjectID<T>(records: Array<T & {_id?: mongoose.Types.ObjectId}>): void
 {
   for (const obj of records)
   {
     delete obj._id;
   }
-  return records as T[];
 }
